rename sdf_utils.py to lift_utils.py
